#20 refactor: 약관 엔티티, 예외, 유틸 코드 공통 모듈로 이동
- 약관 관련 엔티티 및 Enum을 common 모듈로 이동
- 예외, 유틸 클래스 등 공통 코드 common 모듈로 이동
- 모듈 이동에 따른 참조 변경 및 종속성 정리
- scanBasePackages 로 패키지 스캔 범위 조정
- 사용하지 않는 엔티티 TermDetail 삭제
