+++ yeogi-common/src/main/java/com/onerty/yeogi/common/YeogiCommonApplication.java
@@ -2,8 +2,10 @@ package com.onerty.yeogi.common;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class YeogiCommonApplication {
 
     public static void main(String[] args) {
