+++ yeogi-common/src/main/java/com/onerty/yeogi/common/term/TermTitle.java
@@ -1,4 +1,4 @@
-package com.onerty.yeogi.customer.term;
+package com.onerty.yeogi.common.term;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
