+++ yeogi-common/src/main/java/com/onerty/yeogi/common/util/MessageResponse.java
@@ -1,4 +1,4 @@
-package com.onerty.yeogi.customer.util;
+package com.onerty.yeogi.common.util;
 
 public record MessageResponse(String message) {
 }
