+++ yeogi-customer/src/main/java/com/onerty/yeogi/customer/user/UserController.java
@@ -1,13 +1,13 @@
 package com.onerty.yeogi.customer.user;
 
 
+import com.onerty.yeogi.common.util.BaseResponse;
+import com.onerty.yeogi.common.util.MessageResponse;
 import com.onerty.yeogi.customer.term.dto.TermResponse;
 import com.onerty.yeogi.customer.user.dto.NicknameResponse;
 import com.onerty.yeogi.customer.user.dto.UserSignupRequest;
 import com.onerty.yeogi.customer.user.dto.UserSignupResponse;
 import com.onerty.yeogi.customer.user.dto.VerifyCertificationRequest;
-import com.onerty.yeogi.customer.util.BaseResponse;
-import com.onerty.yeogi.customer.util.MessageResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
